Move drag calculation to a new file
